Extract open and models CLI commands
