+++ edsl/__main__.py
@@ -6,9 +6,6 @@ Entry point: edsl = "edsl.__main__:main" (pyproject.toml)
 
 import sys
 import json
-import html
-import tempfile
-import webbrowser
 from typing import Optional
 from pathlib import Path
 
@@ -17,7 +14,9 @@ import click
 from edsl.cli_commands import auth as auth_commands
 from edsl.cli_commands import humanize as humanize_commands
 from edsl.cli_commands import jobs as jobs_commands
+from edsl.cli_commands import models as models_commands
 from edsl.cli_commands import objects as objects_commands
+from edsl.cli_commands import open as open_commands
 from edsl.cli_commands import results as results_commands
 from edsl.cli_commands import run as run_commands
 from edsl.cli_commands import schema as schema_commands
@@ -30,10 +29,8 @@ from edsl.cli_shared import (
     EXIT_VALIDATION,
     error as _error,
     jsonable as _jsonable,
-    load_openable_json as _load_openable_json,
     output as _output,
     read_json_file as _read_json_file,
-    read_package_manifest as _read_package_manifest,
 )
 
 
@@ -147,7 +144,9 @@ def humanize(ctx):
 auth_commands.register(app, auth)
 humanize_commands.register(humanize)
 jobs_commands.register(jobs)
+models_commands.register(app)
 objects_commands.register(app)
+open_commands.register(app)
 results_commands.register(results)
 run_commands.register(app)
 schema_commands.register(schema)
@@ -183,271 +182,6 @@ def _redact_config(config: dict) -> dict:
         else:
             redacted[key] = value
     return redacted
-
-
-# ---------------------------------------------------------------------------
-# edsl open
-# ---------------------------------------------------------------------------
-
-@app.command("open")
-@click.argument("path", type=click.Path(exists=True))
-@click.option("--output", "-o", "output_path", default=None, help="Path to write the generated HTML.")
-@click.option("--browser/--no-browser", default=True, help="Open the generated HTML in a browser.")
-def open_object(path, output_path, browser):
-    """Open an EDSL object as an HTML artifact in a browser."""
-    source_path = Path(path)
-    html_path = _html_output_path(source_path, output_path)
-
-    try:
-        obj = _load_openable_object(source_path)
-        _write_object_html(obj, html_path)
-    except SystemExit:
-        raise
-    except Exception as e:
-        _error(
-            "OPEN_ERROR",
-            f"Failed to render {path} as HTML: {e}",
-            suggestion="Use a saved Survey, AgentList, or Jobs JSON file, or a supported .ep package.",
-            exit_code=EXIT_ERROR,
-        )
-
-    url = html_path.resolve().as_uri()
-    opened = False
-    if browser:
-        try:
-            opened = bool(webbrowser.open(url))
-        except Exception as e:
-            _error(
-                "BROWSER_ERROR",
-                f"Failed to open browser: {e}",
-                suggestion=f"Open this file manually: {html_path}",
-                exit_code=EXIT_ERROR,
-            )
-
-    _output({
-        "source": str(source_path),
-        "html_path": str(html_path),
-        "url": url,
-        "opened": opened,
-        "object_type": _openable_object_type(obj),
-    })
-
-
-def _html_output_path(source_path: Path, output_path: Optional[str]) -> Path:
-    if output_path:
-        return Path(output_path)
-    safe_stem = source_path.name.replace("/", "_")
-    handle = tempfile.NamedTemporaryFile(
-        "w",
-        delete=False,
-        prefix=f"edsl-{safe_stem}-",
-        suffix=".html",
-    )
-    handle.close()
-    return Path(handle.name)
-
-
-def _load_openable_object(path: Path):
-    if path.is_dir():
-        return _load_openable_package(path)
-    if path.suffix == ".ep":
-        return _load_openable_package(path)
-    return _load_openable_json(path)
-
-
-def _load_openable_package(path: Path):
-    manifest = _read_package_manifest(path)
-    class_name = manifest.get("edsl_class_name") or manifest.get("object_type")
-    if class_name == "Survey":
-        from edsl.surveys import Survey
-
-        return Survey.git.open(path)
-    if class_name == "AgentList":
-        from edsl.agents import AgentList
-
-        return AgentList.git.open(path)
-    if class_name == "Jobs":
-        from edsl.jobs import Jobs
-
-        return Jobs.git.open(path)
-    if class_name == "Results":
-        from edsl.results import Results
-
-        return Results.git.open(path)
-    if class_name == "ScenarioList":
-        from edsl.scenarios import ScenarioList
-
-        return ScenarioList.git.open(path)
-    if class_name == "ModelList":
-        from edsl.language_models import ModelList
-
-        return ModelList.git.open(path)
-    _error(
-        "UNSUPPORTED_OBJECT",
-        f"Object package type does not support HTML rendering: {class_name or 'unknown'}",
-        suggestion="Currently supported package types: Survey, AgentList, Jobs, Results, ScenarioList, ModelList.",
-        exit_code=EXIT_USAGE,
-    )
-
-
-def _write_object_html(obj, html_path: Path) -> None:
-    html_path.parent.mkdir(parents=True, exist_ok=True)
-    renderer = getattr(obj, "to_html", None) or getattr(obj, "html", None)
-    if renderer is not None:
-        renderer(filename=str(html_path))
-    else:
-        repr_html = getattr(obj, "_repr_html_", None)
-        if repr_html is not None:
-            body = repr_html()
-        else:
-            to_dataset = getattr(obj, "to_dataset", None)
-            dataset = to_dataset() if to_dataset is not None else None
-            body = dataset._repr_html_() if hasattr(dataset, "_repr_html_") else None
-        if body is None:
-            _error(
-                "UNSUPPORTED_OBJECT",
-                f"Object type does not support HTML rendering: {_openable_object_type(obj)}",
-                exit_code=EXIT_USAGE,
-            )
-        html_path.write_text(
-            _standalone_html_document(
-                title=f"EDSL {_openable_object_type(obj)}",
-                body=body,
-            ),
-            encoding="utf-8",
-        )
-    if not html_path.exists():
-        _error(
-            "OPEN_ERROR",
-            f"HTML renderer did not create expected file: {html_path}",
-            exit_code=EXIT_ERROR,
-        )
-
-
-def _standalone_html_document(title: str, body: str) -> str:
-    return (
-        "<!doctype html>\n"
-        "<html>\n"
-        "<head>\n"
-        '  <meta charset="utf-8">\n'
-        f"  <title>{html.escape(title)}</title>\n"
-        "</head>\n"
-        "<body>\n"
-        f"{body}\n"
-        "</body>\n"
-        "</html>\n"
-    )
-
-
-def _openable_object_type(obj) -> str:
-    return getattr(obj, "object_type", type(obj).__name__)
-
-
-# ---------------------------------------------------------------------------
-# edsl models
-# ---------------------------------------------------------------------------
-
-@app.command()
-@click.option("--service", default=None, help="Filter by service name.")
-@click.option("--search", default=None, help="Wildcard search pattern.")
-@click.option("--text/--no-text", "works_with_text", default=None, help="Filter by text capability.")
-@click.option("--vision/--no-vision", "works_with_images", default=None, help="Filter by image/vision capability.")
-def models(service, search, works_with_text, works_with_images):
-    """List available models."""
-    from edsl.language_models import Model
-
-    # Determine which services have configured keys
-    try:
-        key_info = Model.key_info()
-        configured_services = set()
-        for entry in key_info:
-            if hasattr(entry, 'get'):
-                if entry.get('api_key_set'):
-                    configured_services.add(entry.get('service_name', ''))
-            elif hasattr(entry, 'api_key_set'):
-                if entry.api_key_set:
-                    configured_services.add(getattr(entry, 'service_name', ''))
-    except Exception:
-        configured_services = set()
-
-    warnings = []
-    source = "expected_parrot"
-    try:
-        from edsl.coop import Coop
-
-        available = Coop().fetch_working_models()
-        model_list = []
-        for item in available:
-            model_name = item.get("model")
-            service_name = item.get("service")
-            if service and service_name != service:
-                continue
-            if search and search.lower() not in str(model_name).lower():
-                continue
-            if works_with_text is not None and item.get("works_with_text") is not works_with_text:
-                continue
-            if works_with_images is not None and item.get("works_with_images") is not works_with_images:
-                continue
-            model_list.append({
-                "model_name": model_name,
-                "service_name": service_name,
-                "configured": service_name in configured_services,
-                "works_with_text": item.get("works_with_text"),
-                "works_with_images": item.get("works_with_images"),
-                "usd_per_1M_input_tokens": item.get("usd_per_1M_input_tokens"),
-                "usd_per_1M_output_tokens": item.get("usd_per_1M_output_tokens"),
-            })
-    except Exception as remote_error:
-        if works_with_text is not None or works_with_images is not None:
-            _error(
-                "MODEL_LIST_ERROR",
-                f"Could not fetch model capabilities from Expected Parrot: {remote_error}",
-                suggestion="Retry without --text/--no-text/--vision/--no-vision, or check your network/API key.",
-                exit_code=EXIT_REMOTE,
-            )
-        try:
-            available = Model.available(
-                search_term=search or None,
-                service_name=service or None,
-                local_only=True,
-            )
-        except Exception as e:
-            _error("MODEL_LIST_ERROR", str(e))
-
-        warnings.append(
-            f"Could not fetch models from Expected Parrot; returned local models only: {remote_error}"
-        )
-        source = "local"
-        model_list = []
-        for m in available:
-            model_name = m.model if hasattr(m, 'model') else str(m)
-            service_name = getattr(m, '_inference_service_', '') or getattr(m, 'inference_service', '') or ""
-            model_list.append({
-                "model_name": model_name,
-                "service_name": service_name,
-                "configured": service_name in configured_services,
-                "works_with_text": None,
-                "works_with_images": None,
-                "usd_per_1M_input_tokens": None,
-                "usd_per_1M_output_tokens": None,
-            })
-
-    # Sort alphabetically by service then model
-    model_list.sort(key=lambda x: (x["service_name"], x["model_name"]))
-    _output(
-        {
-            "models": model_list,
-            "source": source,
-            "filters": {
-                "service": service,
-                "search": search,
-                "text": works_with_text,
-                "vision": works_with_images,
-            },
-            "count": len(model_list),
-        },
-        warnings=warnings,
-    )
